MYR-000: renamed component to pascal case
