+++ src/components-v2/Profile/ProfileHeader/ProfileHeaderContainer.tsx
@@ -1,7 +1,7 @@
 import React, {useEffect, useState} from 'react';
 import {useSelector} from 'react-redux';
 
-import {ProfileHeaderComponent} from './';
+import {ProfileHeaderComponent} from '.';
 
 import {useFriendHook} from 'src/components/profile/use-profile-friend.hook';
 import {RootState} from 'src/reducers';
